+++ app/services/auth/fetch_user_service.rb
@@ -4,24 +4,25 @@ module Auth
   class FetchUserService
     prepend ApplicationService
 
+    attr_reader :session
+
     def call(token:)
       @payload = extract_uuid(token)
+      @session = find_session
 
-      fail!('Forbidden') if @payload.blank? || session.blank?
+      fail!('Forbidden') if @payload.blank? || @session.blank?
 
-      @result = session&.user
+      @result = @session&.user
     end
 
     private
 
     def extract_uuid(token)
       JwtEncoder.decode(token)
-    rescue JWT::DecodeError
-      {}
     end
 
-    def session
-      @session ||= Users::Session.where(uuid: @payload.fetch('uuid', '')).first
+    def find_session
+      Users::Session.where(uuid: @payload.fetch('uuid', '')).first
     end
   end
 end