+++ app/services/auth/generate_token_service.rb
@@ -5,10 +5,7 @@ module Auth
     prepend ApplicationService
 
     def call(user:)
-      ActiveRecord::Base.transaction do
-        Users::Session.where(user: user).destroy_all
-        @result = JwtEncoder.encode(uuid: Users::Session.create!(user: user).uuid)
-      end
+      @result = JwtEncoder.encode(uuid: Users::Session.create!(user: user).uuid)
     end
   end
 end