+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - login through Gitlab
+- expiration time to auth tokens
+- removing expired users sessions
 
 ### Modified
 - default insight fields